requirements.txt
pandas >= 2.1
pyomo >= 6.9.0
gamspy >= 1.9.0
pyyaml >= 6.0.2
cerberus >= 1.3.7
highspy >= 1.10.0